+++ tests/testthat/test-dataframe.R
@@ -906,3 +906,24 @@ test_that("pivot args works", {
     )
   )
 })
+
+test_that("rename", {
+  df = pl$DataFrame(mtcars)
+
+  # renaming succeeded
+  a = df$rename(miles_per_gallon = "mpg", horsepower = "hp")$columns
+  expect_false("hp" %in% a)
+  expect_false("mpg" %in% a)
+  expect_true("miles_per_gallon" %in% a)
+  expect_true("horsepower" %in% a)
+
+  # no args are allowed, but does nothing
+  expect_identical(
+    df$rename()$to_list(),
+    df$to_list()
+  )
+
+  # wrapped args in list is equivalent
+  b = df$rename(list(miles_per_gallon = "mpg", horsepower = "hp"))$columns
+  expect_identical(a, b)
+})