+++ tests/testthat/test-lazy.R
@@ -532,3 +532,24 @@ test_that("melt vs data.table::melt", {
     as.list(melt_mod(dtt, id.vars = c("a", "b"), value_vars = c("c")))
   ))
 })
+
+test_that("rename", {
+  lf = pl$DataFrame(mtcars)$lazy()
+
+  # renaming succeeded
+  a = lf$rename(miles_per_gallon = "mpg", horsepower = "hp")$collect()$columns
+  expect_false("hp" %in% a)
+  expect_false("mpg" %in% a)
+  expect_true("miles_per_gallon" %in% a)
+  expect_true("horsepower" %in% a)
+
+  # no args are allowed, but does nothing
+  expect_identical(
+    lf$rename()$collect()$to_list(),
+    lf$collect()$to_list()
+  )
+
+  # wrapped args in list is equivalent
+  b = lf$rename(list(miles_per_gallon = "mpg", horsepower = "hp"))$collect()$columns
+  expect_identical(a, b)
+})