+++ NEWS.md
@@ -1,6 +1,7 @@
 # polars (development version)
 ## What's changed
 
+- New function `rename()` for `LazyFrame` and `DataFrame` (#239)
 - `<DataFrame>$unique()` and `<LazyFrame>$unique()` gain a `maintain_order` argument (#238).
 
 # polars 0.6.1